% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-tic-ui.R
\name{as.character.tic_ui}
\alias{as.character.tic_ui}
\title{Create a character representation of a \code{tic_ui} object}
\usage{
\method{as.character}{tic_ui}(x, depth = 0L, ...)
}
\arguments{
\item{x}{tic_ui object}

\item{depth}{recusive depth}

\item{...}{ignored}
}
\description{
Create a character representation of a \code{tic_ui} object
}
