% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tk-canvas.R
\name{canvas_arc}
\alias{canvas_arc}
\title{Draw an arc on a canvas}
\usage{
canvas_arc(
  canvas,
  x1,
  y1,
  x2,
  y2,
  start,
  extent,
  style,
  fill,
  outline,
  width,
  dash,
  ...
)
}
\arguments{
\item{canvas}{a \code{tic_ui} 'canvas' element.}

\item{x1, y1, x2, y2}{he coordinates of two diagonally opposite corners of a 
rectangular region enclosing the oval that defines the arc}

\item{start}{starting angle of arc in degrees measured counter-clockwise 
from the "3 o'clock" position. Value in range [-360, 360]}

\item{extent}{Size of te angle range occupied by the arc. Value in range [-360, 360]}

\item{style}{how to draw the arc. One of: pieslice (default), chord, fill
\describe{
  \item{\code{pieslice}}{the arc's region is defined by a section of the 
        oval's perimeter plus two line segments, one between the center
         of the oval and each end of the perimeter section}
  \item{\code{chord}}{the arc's region is defined by a section of the 
        oval's perimeter plus a single line segment connecting the two 
        end points of the perimeter section}
  \item{\code{arc}}{the arc's region consists of a section of the perimeter 
        alone. In thiscase the \code{fill} option is ignored. }
}}

\item{fill}{line colour}

\item{outline}{outline colour}

\item{width}{line width e.g. \code{width = 2}}

\item{dash}{Specifies the line's dash pattern. This should be a numeric
vector with alternating lengths of "dash" and "space-between-dash".
E.g. \code{dash = c(6, 4, 2, 4)} produces a dotted-dashed line}

\item{...}{other line creation options. See 
\url{https://www.tcl.tk/man/tcl8.6/TkCmd/canvas.html#M26}}
}
\description{
Draw an arc on a canvas
}
\seealso{
Other canvas: 
\code{\link{canvas_clear}()},
\code{\link{canvas_image}()},
\code{\link{canvas_line}()},
\code{\link{canvas_oval}()},
\code{\link{canvas_plot}()},
\code{\link{canvas_polygon}()},
\code{\link{canvas_rect}()},
\code{\link{canvas_save}()},
\code{\link{canvas_text}()},
\code{\link{tic_canvas}()}
}
\concept{canvas}
