% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tk-canvas.R
\name{canvas_clear}
\alias{canvas_clear}
\title{Clear the canvas of all objects}
\usage{
canvas_clear(canvas)
}
\arguments{
\item{canvas}{a \code{tic_ui} 'canvas' element.}
}
\description{
Clear the canvas of all objects
}
\seealso{
Other canvas: 
\code{\link{canvas_arc}()},
\code{\link{canvas_image}()},
\code{\link{canvas_line}()},
\code{\link{canvas_oval}()},
\code{\link{canvas_plot}()},
\code{\link{canvas_polygon}()},
\code{\link{canvas_rect}()},
\code{\link{canvas_save}()},
\code{\link{canvas_text}()},
\code{\link{tic_canvas}()}
}
\concept{canvas}
