% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tk-canvas.R
\name{canvas_image}
\alias{canvas_image}
\title{Render an image a canvas}
\usage{
canvas_image(canvas, x, y, image, anchor, ...)
}
\arguments{
\item{canvas}{a \code{tic_ui} 'canvas' element.}

\item{x, y}{coorindates for image}

\item{image}{as loaded by \code{load_tkimage()} or otherwise manually created 
with \code{tcltk}}

\item{anchor}{anchor point within image for positioning. default: center
Possible values: n, s, e, w, ne, nw, se, sw, center}

\item{...}{other line creation options. See 
\url{https://www.tcl.tk/man/tcl8.6/TkCmd/canvas.html#M26}}
}
\description{
Render an image a canvas
}
\seealso{
Other canvas: 
\code{\link{canvas_arc}()},
\code{\link{canvas_clear}()},
\code{\link{canvas_line}()},
\code{\link{canvas_oval}()},
\code{\link{canvas_plot}()},
\code{\link{canvas_polygon}()},
\code{\link{canvas_rect}()},
\code{\link{canvas_save}()},
\code{\link{canvas_text}()},
\code{\link{tic_canvas}()}
}
\concept{canvas}
