% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tk-canvas.R
\name{canvas_oval}
\alias{canvas_oval}
\title{Draw an oval on a canvas}
\usage{
canvas_oval(canvas, x1, y1, x2, y2, fill, outline, width, dash, ...)
}
\arguments{
\item{canvas}{a \code{tic_ui} 'canvas' element.}

\item{x1, y1, x2, y2}{corners of rectangle enclosing the oval}

\item{fill}{line colour}

\item{outline}{outline colour}

\item{width}{line width e.g. \code{width = 2}}

\item{dash}{Specifies the line's dash pattern. This should be a numeric
vector with alternating lengths of "dash" and "space-between-dash".
E.g. \code{dash = c(6, 4, 2, 4)} produces a dotted-dashed line}

\item{...}{other line creation options. See 
\url{https://www.tcl.tk/man/tcl8.6/TkCmd/canvas.html#M26}}
}
\description{
Draw an oval on a canvas
}
\seealso{
Other canvas: 
\code{\link{canvas_arc}()},
\code{\link{canvas_clear}()},
\code{\link{canvas_image}()},
\code{\link{canvas_line}()},
\code{\link{canvas_plot}()},
\code{\link{canvas_polygon}()},
\code{\link{canvas_rect}()},
\code{\link{canvas_save}()},
\code{\link{canvas_text}()},
\code{\link{tic_canvas}()}
}
\concept{canvas}
