% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tk-canvas.R
\name{canvas_plot}
\alias{canvas_plot}
\title{Render a plot to the canvas}
\usage{
canvas_plot(canvas, plot, width, height, x = 0, y = 0, anchor = "nw")
}
\arguments{
\item{canvas}{a \code{tic_ui} 'canvas' element.}

\item{plot}{object to be plotted.  Anything support by \code{ggplot2::ggsave()}}

\item{width, height}{size of plot in pizels}

\item{x, y}{coorindates for image. default (0, 0)}

\item{anchor}{anchor point within image for positioning. default: nw
Possible values: n, s, e, w, ne, nw, se, sw, center}
}
\description{
Render a plot to the canvas
}
\seealso{
Other canvas: 
\code{\link{canvas_arc}()},
\code{\link{canvas_clear}()},
\code{\link{canvas_image}()},
\code{\link{canvas_line}()},
\code{\link{canvas_oval}()},
\code{\link{canvas_polygon}()},
\code{\link{canvas_rect}()},
\code{\link{canvas_save}()},
\code{\link{canvas_text}()},
\code{\link{tic_canvas}()}
}
\concept{canvas}
