% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tk-canvas.R
\name{canvas_polygon}
\alias{canvas_polygon}
\title{Draw a polygon on a canvas}
\usage{
canvas_polygon(
  canvas,
  xs,
  ys,
  fill,
  outline,
  width,
  smooth,
  joinstyle,
  dash,
  ...
)
}
\arguments{
\item{canvas}{a \code{tic_ui} 'canvas' element.}

\item{xs}{vectors of coordinates}

\item{ys}{vectors of coordinates}

\item{fill}{line colour}

\item{outline}{outline colour}

\item{width}{line width e.g. \code{width = 2}}

\item{smooth}{should the line be draw as quadratic beziers instead of line segements?
logical. default: FALSE}

\item{joinstyle}{Specifies the ways in which joints are to be drawn at 
the vertices of the line. Possible values: bevel, miter, or round). 
If this option is not specified then it defaults to round.}

\item{dash}{Specifies the line's dash pattern. This should be a numeric
vector with alternating lengths of "dash" and "space-between-dash".
E.g. \code{dash = c(6, 4, 2, 4)} produces a dotted-dashed line}

\item{...}{other line creation options. See 
\url{https://www.tcl.tk/man/tcl8.6/TkCmd/canvas.html#M26}}
}
\description{
Draw a polygon on a canvas
}
\seealso{
Other canvas: 
\code{\link{canvas_arc}()},
\code{\link{canvas_clear}()},
\code{\link{canvas_image}()},
\code{\link{canvas_line}()},
\code{\link{canvas_oval}()},
\code{\link{canvas_plot}()},
\code{\link{canvas_rect}()},
\code{\link{canvas_save}()},
\code{\link{canvas_text}()},
\code{\link{tic_canvas}()}
}
\concept{canvas}
