% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tk-canvas.R
\name{canvas_save}
\alias{canvas_save}
\title{Save the contents of the canvas to file}
\usage{
canvas_save(canvas, filename, ps_density = NULL, ...)
}
\arguments{
\item{canvas}{a \code{tic_ui} 'canvas' element.}

\item{filename}{where the image should be saved.  The image suffix will be 
used as the \code{format} argument to the call to \code{magick::image_write()}}

\item{ps_density}{DPI. Resolution to postscript document. This corresponds to 
the \code{density} argument in {magick::image_read()}.  Default: NULL
means to use whatever \code{magick} package defaults to.}

\item{...}{all other arguments passed to \code{magick::image_write()}}
}
\description{
tcl/tk only exports the canvas as a PS (Postscript) file.  In this function,
the \code{magick} package is used to read the postscript file and render it
to an image file based upon the suffix supplied in the \code{filename} argument.
}
\seealso{
Other canvas: 
\code{\link{canvas_arc}()},
\code{\link{canvas_clear}()},
\code{\link{canvas_image}()},
\code{\link{canvas_line}()},
\code{\link{canvas_oval}()},
\code{\link{canvas_plot}()},
\code{\link{canvas_polygon}()},
\code{\link{canvas_rect}()},
\code{\link{canvas_text}()},
\code{\link{tic_canvas}()}
}
\concept{canvas}
