% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tk-canvas.R
\name{canvas_text}
\alias{canvas_text}
\title{Draw text on a canvas}
\usage{
canvas_text(canvas, x, y, text, fill, justify, angle, ...)
}
\arguments{
\item{canvas}{a \code{tic_ui} 'canvas' element.}

\item{x, y}{coorindates for text}

\item{text}{string to display}

\item{fill}{text colour}

\item{justify}{one of 'left', 'right', 'center'.  Default: left}

\item{angle}{default: 0}

\item{...}{other line creation options. See 
\url{https://www.tcl.tk/man/tcl8.6/TkCmd/canvas.html#M26}}
}
\description{
Draw text on a canvas
}
\seealso{
Other canvas: 
\code{\link{canvas_arc}()},
\code{\link{canvas_clear}()},
\code{\link{canvas_image}()},
\code{\link{canvas_line}()},
\code{\link{canvas_oval}()},
\code{\link{canvas_plot}()},
\code{\link{canvas_polygon}()},
\code{\link{canvas_rect}()},
\code{\link{canvas_save}()},
\code{\link{tic_canvas}()}
}
\concept{canvas}
