% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-tic-ui.R
\name{.DollarNames.tic_ui}
\alias{.DollarNames.tic_ui}
\title{Autocomplete helper for \code{tic_ui} objects}
\usage{
\method{.DollarNames}{tic_ui}(x, pattern)
}
\arguments{
\item{x}{object}

\item{pattern}{current pattern to match}
}
\description{
This autocomplete helper removes 'type' and 'env' from the suggestions,
and moves the suggestion of 'ID' to last in the list.
}
\details{
This is because the user is most often going to be using this object
to havigate the element tree, and will rarely need to access anything
other than the named child nodes.
}
