% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactive.R
\name{format.reactive_value}
\alias{format.reactive_value}
\alias{format.reactive_textbox}
\alias{print.reactive_value}
\title{Format/print a reactive value}
\usage{
\method{format}{reactive_value}(x, ...)

\method{format}{reactive_textbox}(x, ...)

\method{print}{reactive_value}(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{ignored}
}
\description{
Format/print a reactive value
}
