% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image.R
\name{handle_images}
\alias{handle_images}
\title{Sanitize an argument list to replace any image paths with a tkimage object}
\usage{
handle_images(args)
}
\arguments{
\item{args}{list of args as part of a \code{tic_spec}}
}
\value{
list of args with any \code{image} arguments replaced by a \code{tkimage} object
}
\description{
Sanitize an argument list to replace any image paths with a tkimage object
}
