% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-idle.R
\name{launch_idle_func}
\alias{launch_idle_func}
\title{Setup an idle callback to a user function at the given FPS while the window is alive}
\usage{
launch_idle_func(win, user_func, fps = 30, initial_delay = 100)
}
\arguments{
\item{win}{top level window object}

\item{user_func}{users R function.  This function will be called within the
idle loop within any argumnets}

\item{fps}{desired frame rate. default: 30}

\item{initial_delay}{initial delay before running function for the first time. 
In milliseconds.  Default: 1000}
}
\description{
To stop the function: 
\itemize{
  \item{Close the window it is attached to}
  \item{Use a global logical value that is consulted inside the user func to determine if any action should be taken}
}
}
