% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{list_font_families}
\alias{list_font_families}
\title{List all the font families present}
\usage{
list_font_families()
}
\value{
character vector of names
}
\description{
List all the font families present
}
