% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-pack.R
\name{pack_opts}
\alias{pack_opts}
\title{Create a list of pack options used during widget creation}
\usage{
pack_opts(anchor, expand, fill, ipadx, ipady, padx, pady, side, ...)
}
\arguments{
\item{anchor}{Specify where to position content within its parent. 
Defaults to 'center'.  The alternative is to specify a string 
consisting only of the letters 'n', 's', 'e', 'w', to indicate
the compass direction to anchor to.  e.g. 'w', or 'sw' for 
'west' (left) or 'southwest' respectively.}

\item{expand}{Should the content should be expanded to consume extra 
space in its parent container? Default: TRUE}

\item{fill}{If parent size is larger than its requested dimensions, 
       this option may be used to stretch the content. 
\describe{
\item{\code{none}}{No stretching of widget}
\item{\code{x}}{Stretch the content horizontally to fill the space}
\item{\code{y}}{Stretch the content vertically to fill the space}
\item{\code{both}}{Stretch the content horizontally and vertically to fill the space}
}}

\item{ipadx, ipady}{How much horizontal/vertical \emph{internal} padding
to leave on the side of each element.  

If you are familiar with HTML/CSSThis is analogous to 
CSS \code{margin}.}

\item{padx, pady}{How much horizontal/vertical \emph{external} padding
to leave on the side of each element. This may be two values in 
order to specify different padding for left vs right, or top 
vs bottom.

If you are familiar with HTML/CSSThis is analogous to 
CSS \code{margin}.}

\item{side}{Which side of the container the content will be packed against?
Possible values: left, right, top, bottom}

\item{...}{extra named args used by the packing spec.}
}
\description{
These pack options specify how a widget is packed into its parent element.
}
\section{tcl/tk}{


See tcl/tk documentation for more information on the packing specification 
\url{https://www.tcl.tk/man/tcl8.6/TkCmd/pack.html}
}

