% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tk-popups.R
\name{popup_color_picker}
\alias{popup_color_picker}
\title{Create a popup window for choosing a colour}
\usage{
popup_color_picker(title = "Choose colourr", ...)
}
\arguments{
\item{title}{Title to display on popup window}

\item{...}{other named arguments used to initialise this widget}
}
\description{
Create a popup window for choosing a colour
}
\section{tcl/tk}{


See tcl/tk documentation for more information on this element 
\url{https://www.tcl.tk/man/tcl8.6/TkCmd/chooseColor.html}
}

\examples{
\dontrun{
popup_color_picker(message = 'hello')
}

}
