% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tk-popups.R
\name{popup_messagebox}
\alias{popup_messagebox}
\title{Display a popup message box}
\usage{
popup_messagebox(message, type, title, ...)
}
\arguments{
\item{message}{message to display. Required.}

\item{type}{one of 'abortretryignore', 'ok', 'okcancel', 'retrycancel',
'yesno', 'yesnocancel'}

\item{title}{window title}

\item{...}{other options (experts)}
}
\description{
Display a popup message box
}
\section{tcl/tk}{


See tcl/tk documentation for more information on this element 
\url{https://www.tcl.tk/man/tcl8.6/TkCmd/messageBox.html}
}

\examples{
\dontrun{
popup_messagebox(message = 'hello')
}
}
