% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tk-popups.R
\name{popup_open_file}
\alias{popup_open_file}
\alias{popup_save_file}
\title{Dialogs for choosing a file to open and save}
\usage{
popup_open_file(multiple, ...)

popup_save_file(confirmoverwrite, ...)
}
\arguments{
\item{multiple}{allow the user to choose multiple files? default: FALSE}

\item{...}{other arguments. See tcltk documentation for all possible options}

\item{confirmoverwrite}{If user selects a filename which already exists,
then show another popup to confirm overwriting this file. 
Default: TRUE.}
}
\description{
Dialogs for choosing a file to open and save
}
\section{tcl/tk}{


See tcl/tk documentation for more information on this element 
\url{https://www.tcl.tk/man/tcl8.6/TkCmd/getOpenFile.html}
}

