% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-ui-spec.R
\name{print.tic_spec}
\alias{print.tic_spec}
\title{print a tic_spec}
\usage{
\method{print}{tic_spec}(x, ...)
}
\arguments{
\item{x}{\code{tic_spec} object}

\item{...}{ignored}
}
\description{
print a tic_spec
}
