% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-tic-ui.R
\name{print.tic_ui}
\alias{print.tic_ui}
\title{Print a \code{tic_ui} object}
\usage{
\method{print}{tic_ui}(x, ...)
}
\arguments{
\item{x}{\code{tic_ui} object}

\item{...}{ignored}
}
\description{
Print a \code{tic_ui} object
}
