% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactive.R
\name{reactive_lgl}
\alias{reactive_lgl}
\alias{reactive_int}
\alias{reactive_chr}
\alias{reactive_dbl}
\alias{reactive_textbox}
\title{Create a reactive variable}
\usage{
reactive_lgl(value = FALSE)

reactive_int(value = 0L)

reactive_chr(value = "")

reactive_dbl(value = 0)

reactive_textbox(value)
}
\arguments{
\item{value}{initial value.}
}
\description{
Create a reactive variable
}
