% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tk-widgets.R
\name{render_ui}
\alias{render_ui}
\title{Render a complete UI given a spec for a window with child elements}
\usage{
render_ui(spec, parent = NULL)
}
\arguments{
\item{spec}{UI spec as created using \code{tic_window()}, 
\code{tic_button()} etc}

\item{parent}{parent object.  The default 'NULL' value is only allowed
for \code{tic_window()} objects. Otherwise this should be a 
\code{tclObj} holding a reference to a window.
TODO: parent could also be something else done by a render_ui call?}
}
\value{
a nested named list of tcl objects making up the UI.  As a side
        effect of unning this code, a tcl/tk window will be opened and
        rendered to this specficiation.
}
\description{
Render a complete UI given a spec for a window with child elements
}
