% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{set_font_family}
\alias{set_font_family}
\title{Set the font family}
\usage{
set_font_family(body = NULL, headings = body)
}
\arguments{
\item{body}{name of font family for body text, buttons, menus etc}

\item{headings}{name of the font family for the h1-h5 headings for labels.
by default this will be set to the same as the \code{body} font.}
}
\description{
Set the font family
}
