% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{set_font_scale}
\alias{set_font_scale}
\title{Globally scale the fonts in the rendered UI}
\usage{
set_font_scale(scale)
}
\arguments{
\item{scale}{numeric scale factor. default 1.0}
}
\value{
none.
}
\description{
Globally scale the fonts in the rendered UI
}
