% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{set_theme}
\alias{set_theme}
\title{Activate a known theme by name}
\usage{
set_theme(theme_name = "r-sun-valley-light")
}
\arguments{
\item{theme_name}{name of theme. By default this will load the built-in 
theme developed for R.  
Use \code{list_themes()} to see what themes are available.}
}
\value{
none
}
\description{
For advanced users, you can load a theme from a file using
\code{tcltk::tcl('source', theme_file)}, before calling \code{set_theme()}
}
\examples{
{
\dontrun{
list_themes()
set_theme('default')
}
}

}
\concept{themes}
