% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tk-widgets.R
\name{tic_canvas}
\alias{tic_canvas}
\title{Create drawing surface widget}
\usage{
tic_canvas(
  background = "#fafafa",
  scrollbars = FALSE,
  bind = NULL,
  pack = NULL,
  ...
)
}
\arguments{
\item{background}{Background colour. Default: '#fafafa'}

\item{scrollbars}{include scrollbars on the canvas?  Default: FALSE.
This option has not really been tested.  Scrollbars still look funky.}

\item{bind}{bind commands to particular events on this element. This may be 
a single result of \code{bind_opts()} or a list of them for multiple
events.}

\item{pack}{a named list of pack options for how to incorporate this element into its 
parent container.  Default: NULL means to use the standard packing.
See \code{pack_opts()} as a way of creating a valid list of pack options.}

\item{...}{Other arguments are parsed as follows:
 \describe{
   \item{\code{named arguments}}{Further options to be used during the creation 
         of this widget.  See the tcl/tk documentation for all arguments possible
         for this widget.}
   \item{\code{unnamed arguments}}{Container widgets (e.g. \code{tic_frame()}) treat any 
         unnamed arguments as child objects.   Non-container widgets 
         (e.g \code{tic_button()}) will
         raise an error if there are any unnamed widgets.}
}}
}
\value{
handle on the tcl/tk object. TODO: better language needed here.
}
\description{
This command will create a canvas widget as part of UI creation, but 
all the interesting things the user would want to do with a canvas are interactive things
\emph{after} the UI is available to the user.
}
\details{
So to make actual use of canvas at the moment, you'll have to 
be prepared to write some code with the `{tcltk}` package and 
read lots of documentation!
}
\section{tcl/tk}{


See tcl/tk documentation for more information on this element 
\url{https://www.tcl.tk/man/tcl8.6/TkCmd/canvas.html}
}

\seealso{
Other widgets: 
\code{\link{tic_button}()},
\code{\link{tic_checkbutton}()},
\code{\link{tic_combobox}()},
\code{\link{tic_label}()},
\code{\link{tic_menubutton}()},
\code{\link{tic_menuitem}()},
\code{\link{tic_menu}()},
\code{\link{tic_progressbar}()},
\code{\link{tic_radiobutton}()},
\code{\link{tic_separator}()},
\code{\link{tic_sizegrip}()},
\code{\link{tic_slider}()},
\code{\link{tic_spinbox}()},
\code{\link{tic_submenu}()},
\code{\link{tic_textbox}()},
\code{\link{tic_textentry}()}

Other canvas: 
\code{\link{canvas_arc}()},
\code{\link{canvas_clear}()},
\code{\link{canvas_image}()},
\code{\link{canvas_line}()},
\code{\link{canvas_oval}()},
\code{\link{canvas_plot}()},
\code{\link{canvas_polygon}()},
\code{\link{canvas_rect}()},
\code{\link{canvas_save}()},
\code{\link{canvas_text}()}
}
\concept{canvas}
\concept{widgets}
