% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tk-widgets.R
\name{tic_label}
\alias{tic_label}
\title{Create a text label}
\usage{
tic_label(text, textvariable, width, style, bind = NULL, pack = NULL, ...)
}
\arguments{
\item{text}{Specifies a text string to be displayed inside the widget.
\describe{
  \item{\code{text}}{Simple character string containing the name}
  \item{\code{textvariable}}{Reactive variable containing a string}
}}

\item{textvariable}{Specifies a text string to be displayed inside the widget.
\describe{
  \item{\code{text}}{Simple character string containing the name}
  \item{\code{textvariable}}{Reactive variable containing a string}
}}

\item{width}{If greater than zero, specifies how much space, in character 
widths, to allocate for the text label. If less than zero, specifies 
a minimum width. If zero or unspecified, the natural width of the 
text label is used.}

\item{style}{specify the style for this label. Possible values are "h1" 
through "h5" which are analogues to the h1 to h5 headings in HTML, with
"h1" being the largest heading size, down to "h5" being the smallest.}

\item{bind}{bind commands to particular events on this element. This may be 
a single result of \code{bind_opts()} or a list of them for multiple
events.}

\item{pack}{a named list of pack options for how to incorporate this element into its 
parent container.  Default: NULL means to use the standard packing.
See \code{pack_opts()} as a way of creating a valid list of pack options.}

\item{...}{Other arguments are parsed as follows:
 \describe{
   \item{\code{named arguments}}{Further options to be used during the creation 
         of this widget.  See the tcl/tk documentation for all arguments possible
         for this widget.}
   \item{\code{unnamed arguments}}{Container widgets (e.g. \code{tic_frame()}) treat any 
         unnamed arguments as child objects.   Non-container widgets 
         (e.g \code{tic_button()}) will
         raise an error if there are any unnamed widgets.}
}}
}
\value{
handle on the tcl/tk object
}
\description{
Create a text label
}
\section{tcl/tk}{


See tcl/tk documentation for more information on this element 
\url{https://www.tcl.tk/man/tcl8.6/TkCmd/ttk_label.htm}
}

\seealso{
Other widgets: 
\code{\link{tic_button}()},
\code{\link{tic_canvas}()},
\code{\link{tic_checkbutton}()},
\code{\link{tic_combobox}()},
\code{\link{tic_menubutton}()},
\code{\link{tic_menuitem}()},
\code{\link{tic_menu}()},
\code{\link{tic_progressbar}()},
\code{\link{tic_radiobutton}()},
\code{\link{tic_separator}()},
\code{\link{tic_sizegrip}()},
\code{\link{tic_slider}()},
\code{\link{tic_spinbox}()},
\code{\link{tic_submenu}()},
\code{\link{tic_textbox}()},
\code{\link{tic_textentry}()}
}
\concept{widgets}
