% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tk-menu.R
\name{tic_menubutton}
\alias{tic_menubutton}
\title{Button that pops down a menu when pressed}
\usage{
tic_menubutton(..., text, bind = NULL, pack = NULL)
}
\arguments{
\item{...}{Other arguments are parsed as follows:
 \describe{
   \item{\code{named arguments}}{Further options to be used during the creation 
         of this widget.  See the tcl/tk documentation for all arguments possible
         for this widget.}
   \item{\code{unnamed arguments}}{Container widgets (e.g. \code{tic_frame()}) treat any 
         unnamed arguments as child objects.   Non-container widgets 
         (e.g \code{tic_button()}) will
         raise an error if there are any unnamed widgets.}
}}

\item{text}{Specifies a text string to be displayed inside the widget.
\describe{
  \item{\code{text}}{Simple character string containing the name}
  \item{\code{textvariable}}{Reactive variable containing a string}
}}

\item{bind}{bind commands to particular events on this element. This may be 
a single result of \code{bind_opts()} or a list of them for multiple
events.}

\item{pack}{how this object should be packed into its parent.  \code{pack} has
no effect for \code{tic_menu()} or \code{tic_menu_item()}.}
}
\value{
handle on the tcl/tk object
}
\description{
A menubutton widget displays a textual label and/or image, and displays a 
menu when pressed.
}
\details{
Include items in this menu using \code{tic_menuitem()} and \code{tic_submenu()}
}
\section{tcl/tk}{


See tcl/tk documentation for more information on this element 
\url{https://www.tcl.tk/man/tcl8.6/TkCmd/ttk_menubutton.htm}
}

\examples{
tic_menubutton(
  text = "Press for menu",
  tic_menuitem("Run a command", menutype = "command", command = function() {
    message("This is where a command is run")
  }),
  tic_submenu(
    label = "Sub menu here",
    tic_menuitem("Run this", "command", command = function() { message("Hello")})
  )
)


}
\seealso{
Other widgets: 
\code{\link{tic_button}()},
\code{\link{tic_canvas}()},
\code{\link{tic_checkbutton}()},
\code{\link{tic_combobox}()},
\code{\link{tic_label}()},
\code{\link{tic_menuitem}()},
\code{\link{tic_menu}()},
\code{\link{tic_progressbar}()},
\code{\link{tic_radiobutton}()},
\code{\link{tic_separator}()},
\code{\link{tic_sizegrip}()},
\code{\link{tic_slider}()},
\code{\link{tic_spinbox}()},
\code{\link{tic_submenu}()},
\code{\link{tic_textbox}()},
\code{\link{tic_textentry}()}
}
\concept{widgets}
