% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tk-containers.R
\name{tic_notebook}
\alias{tic_notebook}
\title{Notebook:  Multi-paned container window i.e. a window with multiple tabs.}
\usage{
tic_notebook(..., labels, bind = NULL, pack = NULL)
}
\arguments{
\item{...}{Other arguments are parsed as follows:
 \describe{
   \item{\code{named arguments}}{Further options to be used during the creation 
         of this widget.  See the tcl/tk documentation for all arguments possible
         for this widget.}
   \item{\code{unnamed arguments}}{Container widgets (e.g. \code{tic_frame()}) treat any 
         unnamed arguments as child objects.   Non-container widgets 
         (e.g \code{tic_button()}) will
         raise an error if there are any unnamed widgets.}
}}

\item{labels}{Labels to display at the top of each tab. Required.}

\item{bind}{bind commands to particular events on this element. This may be 
a single result of \code{bind_opts()} or a list of them for multiple
events.}

\item{pack}{a named list of pack options for how to incorporate this element into its 
parent container.  Default: NULL means to use the standard packing.
See \code{pack_opts()} as a way of creating a valid list of pack options.}
}
\value{
handle on the tcl/tk object
}
\description{
A notebook widget manages a collection of windows and displays a single 
one at a time.
}
\details{
Each content window is associated with a tab, which the user may select 
to change the currently-displayed window.
}
\section{tcl/tk}{


See tcl/tk documentation for more information on this element 
\url{https://www.tcl.tk/man/tcl8.6/TkCmd/ttk_notebook.htm}
}

