% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tk-containers.R
\name{tic_panedwindow}
\alias{tic_panedwindow}
\title{Panedwindow:  Multi-pane container window}
\usage{
tic_panedwindow(..., sizes = NULL, orient, bind = NULL, pack = NULL)
}
\arguments{
\item{...}{Other arguments are parsed as follows:
 \describe{
   \item{\code{named arguments}}{Further options to be used during the creation 
         of this widget.  See the tcl/tk documentation for all arguments possible
         for this widget.}
   \item{\code{unnamed arguments}}{Container widgets (e.g. \code{tic_frame()}) treat any 
         unnamed arguments as child objects.   Non-container widgets 
         (e.g \code{tic_button()}) will
         raise an error if there are any unnamed widgets.}
}}

\item{sizes}{The relative sizes of the elements in this row or column}

\item{orient}{orientataion. 'horizontal' or 'vertical' (default)}

\item{bind}{bind commands to particular events on this element. This may be 
a single result of \code{bind_opts()} or a list of them for multiple
events.}

\item{pack}{a named list of pack options for how to incorporate this element into its 
parent container.  Default: NULL means to use the standard packing.
See \code{pack_opts()} as a way of creating a valid list of pack options.}
}
\value{
handle on the tcl/tk object
}
\description{
A ttk::panedwindow widget displays a number of subwindows, stacked 
either vertically or horizontally. The user may adjust the relative sizes 
of the subwindows by dragging the sash between panes.
}
\section{tcl/tk}{


See tcl/tk documentation for more information on this element 
\url{https://www.tcl.tk/man/tcl8.6/TkCmd/ttk_panedwindow.htm}
}

