% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tk-widgets.R
\name{tic_radiobutton}
\alias{tic_radiobutton}
\title{Radiobutton: Mutually exclusive option widget}
\usage{
tic_radiobutton(
  text,
  value,
  variable,
  command,
  textvariable,
  width,
  bind = NULL,
  pack = NULL,
  ...
)
}
\arguments{
\item{text}{Specifies a text string to be displayed inside the widget.
\describe{
  \item{\code{text}}{Simple character string containing the name}
  \item{\code{textvariable}}{Reactive variable containing a string}
}}

\item{value}{The value to store in the associated \code{variable} when the widget is selected.}

\item{variable}{The reactive variable holding the state for a set of 
radiobuttons.}

\item{command}{Function to evaluate when the widget is invoked.}

\item{textvariable}{Specifies a text string to be displayed inside the widget.
\describe{
  \item{\code{text}}{Simple character string containing the name}
  \item{\code{textvariable}}{Reactive variable containing a string}
}}

\item{width}{If greater than zero, specifies how much space, in character 
widths, to allocate for the text label. If less than zero, specifies 
a minimum width. If zero or unspecified, the natural width of the 
text label is used.}

\item{bind}{bind commands to particular events on this element. This may be 
a single result of \code{bind_opts()} or a list of them for multiple
events.}

\item{pack}{a named list of pack options for how to incorporate this element into its 
parent container.  Default: NULL means to use the standard packing.
See \code{pack_opts()} as a way of creating a valid list of pack options.}

\item{...}{Other arguments are parsed as follows:
 \describe{
   \item{\code{named arguments}}{Further options to be used during the creation 
         of this widget.  See the tcl/tk documentation for all arguments possible
         for this widget.}
   \item{\code{unnamed arguments}}{Container widgets (e.g. \code{tic_frame()}) treat any 
         unnamed arguments as child objects.   Non-container widgets 
         (e.g \code{tic_button()}) will
         raise an error if there are any unnamed widgets.}
}}
}
\value{
handle on the tcl/tk object
}
\description{
Radiobutton widgets are used in groups to show or change a set of 
mutually-exclusive options.
}
\details{
Radiobuttons are linked to a reactive variable, and have an associated value; 
when a radiobutton is clicked, it sets the variable to its associated value.
}
\section{tcl/tk}{


See tcl/tk documentation for more information on this element 
\url{https://www.tcl.tk/man/tcl8.6/TkCmd/ttk_radiobutton.htm}
}

\examples{
choices <- c('alpha', 'bravo', 'charlie')
state <- reactive_chr(choices[1])
tic_window(
  title = "demo",
  tic_radiobutton(text = choices[1], value = choices[1], variable = state),
  tic_radiobutton(text = choices[2], value = choices[2], variable = state),
  tic_radiobutton(text = choices[3], value = choices[3], variable = state)
)


}
\seealso{
Other widgets: 
\code{\link{tic_button}()},
\code{\link{tic_canvas}()},
\code{\link{tic_checkbutton}()},
\code{\link{tic_combobox}()},
\code{\link{tic_label}()},
\code{\link{tic_menubutton}()},
\code{\link{tic_menuitem}()},
\code{\link{tic_menu}()},
\code{\link{tic_progressbar}()},
\code{\link{tic_separator}()},
\code{\link{tic_sizegrip}()},
\code{\link{tic_slider}()},
\code{\link{tic_spinbox}()},
\code{\link{tic_submenu}()},
\code{\link{tic_textbox}()},
\code{\link{tic_textentry}()}
}
\concept{widgets}
