% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tk-containers.R
\name{tic_row}
\alias{tic_row}
\alias{tic_col}
\title{Create a container element oriented as a row or column of elements}
\usage{
tic_row(..., bind = NULL, pack = NULL, pack_def = NULL)

tic_col(..., bind = NULL, pack = NULL, pack_def = NULL)
}
\arguments{
\item{...}{Other arguments are parsed as follows:
 \describe{
   \item{\code{named arguments}}{Further options to be used during the creation 
         of this widget.  See the tcl/tk documentation for all arguments possible
         for this widget.}
   \item{\code{unnamed arguments}}{Container widgets (e.g. \code{tic_frame()}) treat any 
         unnamed arguments as child objects.   Non-container widgets 
         (e.g \code{tic_button()}) will
         raise an error if there are any unnamed widgets.}
}}

\item{bind}{bind commands to particular events on this element. This may be 
a single result of \code{bind_opts()} or a list of them for multiple
events.}

\item{pack}{a named list of pack options for how to incorporate this element into its 
parent container.  Default: NULL means to use the standard packing.
See \code{pack_opts()} as a way of creating a valid list of pack options.}

\item{pack_def}{Default packing options for children of this object. This
can be overriden by setting \code{pack} explicitly on child elemnets
you want to control packing for.}
}
\value{
handle on the tcl/tk object
}
\description{
This is not a direct analogy for a single tk element
}
\section{tcl/tk}{


This is not a direct implementation for an existing tk element. It is implemented
as a \code{ttk::frame} element with horizontal or vertical packing by default.
}

