% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tk-widgets.R
\name{tic_textentry}
\alias{tic_textentry}
\title{Textentry: Editable text field widget}
\usage{
tic_textentry(
  textvariable,
  validate,
  validatecommand,
  width,
  bind = NULL,
  pack = NULL,
  ...
)
}
\arguments{
\item{textvariable}{Reactive value for the value displayed in this widget}

\item{validate}{when to vdlidate the contents of this text box. Possible 
values: none, key, focus, focusin, focusout, all.  If this is not
set then the \code{validatecommand} function will never be run.}

\item{validatecommand}{function to call when validation event occurs. This
function must return a non-NA boolean value.}

\item{width}{If greater than zero, specifies how much space, in character 
widths, to allocate for the text label. If less than zero, specifies 
a minimum width. If zero or unspecified, the natural width of the 
text label is used.}

\item{bind}{bind commands to particular events on this element. This may be 
a single result of \code{bind_opts()} or a list of them for multiple
events.}

\item{pack}{a named list of pack options for how to incorporate this element into its 
parent container.  Default: NULL means to use the standard packing.
See \code{pack_opts()} as a way of creating a valid list of pack options.}

\item{...}{Other arguments are parsed as follows:
 \describe{
   \item{\code{named arguments}}{Further options to be used during the creation 
         of this widget.  See the tcl/tk documentation for all arguments possible
         for this widget.}
   \item{\code{unnamed arguments}}{Container widgets (e.g. \code{tic_frame()}) treat any 
         unnamed arguments as child objects.   Non-container widgets 
         (e.g \code{tic_button()}) will
         raise an error if there are any unnamed widgets.}
}}
}
\value{
handle on the tcl/tk object
}
\description{
A textentry widget displays a one-line text string and allows that string
to be edited by the user.
}
\details{
The value of the string is linked to a reacive variaable with the \code{textvariable}
argument. 

Note: in tcl/tk this object is known as \code{entry} widget, rather than 
\code{textentry}.
}
\section{tcl/tk}{


See tcl/tk documentation for more information on this element 
\url{https://www.tcl.tk/man/tcl8.6/TkCmd/ttk_entry.htm}
}

\seealso{
Other widgets: 
\code{\link{tic_button}()},
\code{\link{tic_canvas}()},
\code{\link{tic_checkbutton}()},
\code{\link{tic_combobox}()},
\code{\link{tic_label}()},
\code{\link{tic_menubutton}()},
\code{\link{tic_menuitem}()},
\code{\link{tic_menu}()},
\code{\link{tic_progressbar}()},
\code{\link{tic_radiobutton}()},
\code{\link{tic_separator}()},
\code{\link{tic_sizegrip}()},
\code{\link{tic_slider}()},
\code{\link{tic_spinbox}()},
\code{\link{tic_submenu}()},
\code{\link{tic_textbox}()}
}
\concept{widgets}
