% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tk-containers.R
\name{tic_window}
\alias{tic_window}
\title{Create the top level window}
\usage{
tic_window(
  ...,
  title = "{tickle}",
  width = NULL,
  height = NULL,
  bind = NULL,
  pack = NULL,
  pack_def = NULL
)
}
\arguments{
\item{...}{Other arguments are parsed as follows:
 \describe{
   \item{\code{named arguments}}{Further options to be used during the creation 
         of this widget.  See the tcl/tk documentation for all arguments possible
         for this widget.}
   \item{\code{unnamed arguments}}{Container widgets (e.g. \code{tic_frame()}) treat any 
         unnamed arguments as child objects.   Non-container widgets 
         (e.g \code{tic_button()}) will
         raise an error if there are any unnamed widgets.}
}}

\item{title}{Window Title}

\item{width, height}{width and height of window.  If not given then 
UI will be automatically sized.}

\item{bind}{bind commands to particular events on this element. This may be 
a single result of \code{bind_opts()} or a list of them for multiple
events.}

\item{pack}{a named list of pack options for how to incorporate this element into its 
parent container.  Default: NULL means to use the standard packing.
See \code{pack_opts()} as a way of creating a valid list of pack options.}

\item{pack_def}{Default packing options for children of this object. This
can be overriden by setting \code{pack} explicitly on child elemnets
you want to control packing for.}
}
\value{
the tcl/tk window handle
}
\description{
Create the top level window
}
\section{tcl/tk}{


See tcl/tk documentation for more information on this element 
\url{https://www.tcl.tk/man/tcl8.6/TkCmd/toplevel.html}
}

\seealso{
Other widgets containers: 
\code{\link{tic_labelframe}()}
}
\concept{widgets containers}
